refactor: merge BudOptions into TenancyOptions
