+++ src/TenancyOptions.php
@@ -107,4 +107,80 @@ class TenancyOptions
         return self::shouldEnableAllOverrides($tenancy)
                || in_array($service, self::enabledOverrides($tenancy) ?? [], true);
     }
+
+    /**
+     * @param string $store
+     *
+     * @return string[]
+     */
+    public static function useDefaultStore(string $store): array
+    {
+        return [
+            'config:store.default' => $store,
+        ];
+    }
+
+    /**
+     * @param string $store
+     *
+     * @return string[]
+     */
+    public static function alwaysUseStore(string $store): array
+    {
+        return [
+            'config:store.fixed' => $store,
+        ];
+    }
+
+    /**
+     * @param \Sprout\Contracts\Tenancy<*> $tenancy
+     *
+     * @return bool
+     */
+    public static function hasDefaultStore(Tenancy $tenancy): bool
+    {
+        return $tenancy->hasOption('config:store.default');
+    }
+
+    /**
+     * @param \Sprout\Contracts\Tenancy<*> $tenancy
+     *
+     * @return string|null
+     */
+    public static function getDefaultStore(Tenancy $tenancy): ?string
+    {
+        $store = $tenancy->optionConfig('config:store.default');
+
+        if (is_string($store)) {
+            return $store;
+        }
+
+        return null;
+    }
+
+    /**
+     * @param \Sprout\Contracts\Tenancy<*> $tenancy
+     *
+     * @return bool
+     */
+    public static function isLockedToStore(Tenancy $tenancy): bool
+    {
+        return $tenancy->hasOption('config:store.fixed');
+    }
+
+    /**
+     * @param \Sprout\Contracts\Tenancy<*> $tenancy
+     *
+     * @return string|null
+     */
+    public static function getLockedStore(Tenancy $tenancy): ?string
+    {
+        $store = $tenancy->optionConfig('config:store.fixed');
+
+        if (is_string($store)) {
+            return $store;
+        }
+
+        return null;
+    }
 }
